pyproject.toml
[tool.poetry]
name = "autto"
version = "0.1.0"
description = ""
authors = ["skgur07 <pigjaoki0970@gmail.com>"]
readme = "README.md"
packages = [{include = "autto", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
numpy = "^2.2.5"
requests = "^2.32.3"
python-telegram-bot = "^22.1"
schedule = "^1.2.2"
python-dotenv = "^1.1.0"
python-kis = "^2.1.3"
pykrx = "^1.0.48"
mojito2 = "^0.1.6"
xai-sdk = "^0.3.0"
openai = "^1.84.0"
sqlmodel = "^0.0.24"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
